Thanks to DNode all client/server communication is now done through Remote Method Invocation (RMI), which is a wonderful thing :). This applies for realtime editor, the REST API on the front-end is still served through HTTP.
I've also implemented authentication, though it's done through dummy accounts (demo, demo2, etc.) for now. Realtime synchronization now also works for node insertion/deletion/movement operations. Document's that are in progress, are now fetched from the swarm (in a p2p manner) instead of the database (which holds an outdated version at this time).

This was a huge refactoring that solved a lot of issues. We're getting closer now — still not there. ;-)